Moved WorkflowNodeDataFilter from workflow_runner to workflow_filter.py
