+++ studio/app/common/routers/run.py
@@ -9,14 +9,12 @@ from studio.app.common.core.workflow.workflow import (
     NodeItem,
     RunItem,
 )
+from studio.app.common.core.workflow.workflow_filter import WorkflowNodeDataFilter
 from studio.app.common.core.workflow.workflow_result import (
     WorkflowMonitor,
     WorkflowResult,
 )
-from studio.app.common.core.workflow.workflow_runner import (
-    WorkflowNodeDataFilter,
-    WorkflowRunner,
-)
+from studio.app.common.core.workflow.workflow_runner import WorkflowRunner
 from studio.app.common.core.workspace.workspace_dependencies import (
     is_workspace_available,
     is_workspace_owner,
